python-server/requirements.txt
# Core dependencies
<<<<<<< HEAD
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.12.0
httpx==0.24.1

# Database and storage
pymongo==4.6.0
pinecone-client==2.2.2

# Machine Learning and Data Processing
numpy==1.24.3

# LangChain and related
langchain==0.1.0
langchain-community==0.0.1
tiktoken==0.5.1

# Utilities
python-multipart==0.0.6
=======
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
httpx>=0.24.1

# Database and storage
pymongo>=4.6.0
pinecone-client>=2.2.2

# Machine Learning and Data Processing
numpy>=1.24.3

# LangChain and related
langchain>=0.1.0
langchain-community>=0.0.1
tiktoken>=0.5.1

# Utilities
python-multipart==0.0.6
>>>>>>> f0798f5c64f00a718de2cb8dcbed1d20fd9fe46c
